+++ src/app/(auth)/sign-in/page.tsx
@@ -1,6 +1,6 @@
-import AuthCard from "@/components/auth/AuthCard";
-import SignInForm from "@/components/auth/SignInForm";
-import { auth } from "@/lib/auth";
+import AuthCard from "@/modules/auth/components/AuthCard";
+import SignInForm from "@/modules/auth/components/SignInForm";
+import { auth } from "@/modules/auth/lib/auth";
 import { headers } from "next/headers";
 import { redirect } from "next/navigation";
 
@@ -12,7 +12,7 @@ export default async function SignInPage() {
   if (session) {
     redirect("/");
   }
-  
+
   return (
     <AuthCard variant="signin">
       <SignInForm />
