+++ src/app/(auth)/sign-up/page.tsx
@@ -1,6 +1,6 @@
-import AuthCard from "@/components/auth/AuthCard";
-import SignUpForm from "@/components/auth/SignUpForm";
-import { auth } from "@/lib/auth";
+import AuthCard from "@/modules/auth/components/AuthCard";
+import SignUpForm from "@/modules/auth/components/SignUpForm";
+import { auth } from "@/modules/auth/lib/auth";
 import { headers } from "next/headers";
 import { redirect } from "next/navigation";
