+++ src/lib/resend.ts
@@ -1,0 +1,5 @@
+import { Resend } from "resend";
+
+const resend = new Resend(process.env.RESEND_API_KEY);
+
+export default resend;
