+++ src/modules/auth/components/FormDivider.tsx
@@ -1,5 +1,3 @@
-import React from 'react'
-
 export default function FormDivider({ text }: { text: string }) {
   return (
     <div className="after:border-border relative text-center text-sm after:absolute after:inset-0 after:top-1/2 after:z-0 after:flex after:items-center after:border-t">
