+++ src/modules/auth/lib/auth.ts
